requirements.txt
certifi==2019.3.9
chardet==3.0.4
config==0.4.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-bootstrap4==0.0.6
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.7.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
Pillow==5.4.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
pyuploadcare==2.5.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
whitenoise==3.3.1
